Pin envoy versions in examples to v1.9.1
There is no `latest` tag any more, and the Envoy configuration in
the examples is not compatible with the latest Envoy version
(v1.21.2) in any case.

Other, later, versions of Envoy probably also work but given the
timestamps on the file the latest version available at the time
the example was written was either v1.9.1 or v1.10.0, I have
verified v1.9.1 works.